% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{base_Flow.DFM}
\alias{base_Flow.DFM}
\title{Base flow}
\usage{
base_Flow.DFM(
  Streamflow,
  param_filter = 0.925,
  passes_N = 1,
  negativ_step = TRUE
)
}
\arguments{
\item{Streamflow}{[xts]-2D(time, space) Time serious Streamflow in difficult location.}

\item{param_filter}{[num]-range() Filter parameter that enables the shape of the separation to be altered.}

\item{passes_N}{[integer] The number of pass, see [Arnold.1995].}

\item{negativ_step}{[logi] TRUE | FALSE for with or without inveser order, see [Arnold.1995].}
}
\description{
Base flow
}
\examples{
rslt <- base_Flow.DFM(xts_Q)
rslt <- base_Flow.DFM(xts_Q, passes_N = 3, negativ_step = FALSE)
}
