% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{base_Flow.DFM_matrix}
\alias{base_Flow.DFM_matrix}
\title{Base flow}
\usage{
base_Flow.DFM_matrix(Streamflow, param_filter = 0.925)
}
\arguments{
\item{Streamflow}{[xts]-2D(time, space) Time serious Streamflow in difficult location.}

\item{param_filter}{[num]-range() Filter parameter that enables the shape of the separation to be altered.}
}
\value{
baseflow [xts]-2D(time, space) Time serious baseflow in difficult location.
}
\description{
Base flow with digital filter method.
}
\examples{
rslt <- base_Flow.DFM_matrix(xts_Q)
}
\references{
[Arnold.1995]
[eq. 1-2]
}
