% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maginature.R
\name{sig_Autocor}
\alias{sig_Autocor}
\title{Autocorrelation}
\usage{
sig_Autocor(Data, time_lag = 1)
}
\arguments{
\item{Data}{[xts / array]-2D(time, space) Time serious Data in difficult location.}

\item{time_lag}{[integer] The number of lag.}
}
\value{
Autocorrelation [array]-2D(1, space)
}
\description{
Autocorrelation
}
\examples{
rslt <- sig_Autocor(xts_Q)
}
