% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{sig_Baseflow_index.DFM}
\alias{sig_Baseflow_index.DFM}
\title{Baseflow index}
\usage{
sig_Baseflow_index.DFM(
  Streamflow,
  param_filter = 0.925,
  passes_N = 1,
  negativ_step = TRUE,
  t_scale = apply.yearly
)
}
\arguments{
\item{Streamflow}{[xts]-2D(time, space) Time serious Streamflow in difficult location.}

\item{param_filter}{[num]-range() Filter parameter that enables the shape of the separation to be altered.}

\item{passes_N}{[integer] The number of pass, see [Arnold.1995].}

\item{negativ_step}{[logi] TRUE | FALSE for with or without inveser order, see [Arnold.1995].}

\item{t_scale}{[function / NULL] |apply.daily| for daily, |apply.weekly| for weekly, |apply.monthly| for monthly,
|apply.quarterly| for quarterly, |apply.yearly| for yearly, |NULL| for all of the time.}
}
\value{
Baseflow index [xts]-2D(1, space) Time serious baseflow in difficult location.
}
\description{
Baseflow index
}
\examples{
sig_Baseflow_index.DFM(xts_Q, negativ_step = FALSE, passes_N = 3, t_scale = NULL)
sig_Baseflow_index.DFM(xts_Q)
}
