% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maginature.R
\name{sig_Frequency_threshold}
\alias{sig_Frequency_threshold}
\alias{sig_Duration_threshold}
\title{Frequency or Duration by specified threshold}
\usage{
sig_Frequency_threshold(
  Data,
  higher = TRUE,
  threshold = c(1, 3, 9),
  times_mean = TRUE
)

sig_Duration_threshold(
  Data,
  higher = TRUE,
  threshold = c(1, 3, 9),
  times_mean = TRUE,
  dur_type = mean
)
}
\arguments{
\item{Data}{[xts / array]-2D(time, space) Time serious Data in difficult location.}

\item{higher}{[logi] TRUE | FALSE for the value bigger or smaller than the threshold.}

\item{threshold}{[num / vector] one value or a vactor of threshold, If times_mean = TRUE, it should be the times with mean value;
if not it can be the concrete value in the range of \'Data\' and without times the mean value.}

\item{times_mean}{[logi] TRUE | FALSE for use or not use the times with mean value.}

\item{dur_type}{[function] mean | max | min for the stattistical methods for the qualifiede Durations.}
}
\value{
frequency [num]
}
\description{
Frequency or Duration by specified threshold
}
\section{Functions}{
\itemize{
\item \code{sig_Duration_threshold}: Frequency or Duration by specified threshold
}}

\examples{
 rslt <- sig_Frequency_threshold(xts_Q)
rslt <- sig_Frequency_threshold(xts_Q, higher = FALSE, threshold = c(0.1, 1), times_mean = FALSE)
 rslt <- sig_Duration_threshold(xts_Q)
rslt <- sig_Duration_threshold(xts_Q, higher = FALSE, threshold = c(0.1, 1), times_mean = FALSE, dur_type = max)
}
