% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maginature.R
\name{sig_Mean}
\alias{sig_Mean}
\title{Mean}
\usage{
sig_Mean(Data, t_scale = apply.monthly)
}
\arguments{
\item{Data}{[xts]-2D(time, space) Time serious Data in difficult location.}

\item{t_scale}{[function / NULL] |apply.daily| for daily, |apply.weekly| for weekly, |apply.monthly| for monthly,
|apply.quarterly| for quarterly, |apply.yearly| for yearly, |NULL| for all of the time.}
}
\value{
[xts / array]-2D(time, space) Mean of data.
}
\description{
Mean of data, e.g. streamflow, precipitation, temperature and so on.
}
\examples{
rslt <- sig_Mean(xts_Q)
rslt <- sig_Mean(xts_Q, t_scale = NULL)
}
