% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{sig_Peaks_distribution}
\alias{sig_Peaks_distribution}
\title{Slope of distribution of peaks}
\usage{
sig_Peaks_distribution(
  Data,
  decrease_toleranz = 0,
  tile_low = 10,
  tile_high = 50
)
}
\arguments{
\item{Data}{[xts / array]-2D(time, space) Time serious Data in difficult location.}

\item{decrease_toleranz}{[num] Toleranz for the small decrease.}

\item{tile_low, tile_high}{[num]-range(0-100) The title range for slope of the curve.}
}
\value{
sig_Peaks_distribution [num] [1/step]
}
\description{
Slope of distribution of peaks
}
\examples{
rslt <- sig_Peaks_distribution(xts_Q)
}
