% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{sig_Quantile}
\alias{sig_Quantile}
\title{Quantile / Percentile decreasing}
\usage{
sig_Quantile(Data, tile = 95)
}
\arguments{
\item{Data}{[xts / array]-2D(time, space) Time serious Data in difficult location.}

\item{tile}{[num / vector]-range(0-100) The tiles (positions).}
}
\value{
[xts / array]-2D(tile, space) The data in the tiles.
}
\description{
Quantile / Percentile of data in decreasing, e.g. streamflow, precipitation, temperature and so on.
The Q5 is the high flow, Q95 is the low flow.
For the other scale (if you want) can use apply.yearly(Data, sig_Quantile, tile = c(5, 95)) or apply.weekly apply.monthly apply.quarterly apply.yearly.
}
\examples{
rslt <- sig_Quantile(xts_Q)
rslt <- sig_Quantile(xts_Q, tile = c(10, 50, 90))
}
