% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.R
\name{sig_Rising_limb_density}
\alias{sig_Rising_limb_density}
\title{Rising limb density}
\usage{
sig_Rising_limb_density(Data, decrease_toleranz = 0, min_rising_duration = 1)
}
\arguments{
\item{Data}{[xts / array]-2D(time, space) Time serious Data in difficult location.}

\item{decrease_toleranz}{[num] Toleranz for the small decrease.}

\item{min_rising_duration}{[integer] The smallest duration for a decrease.}
}
\value{
sig_Rising_limb_density [num] [1/step]
}
\description{
Rising limb density
}
\examples{
rslt <- sig_Rising_limb_density(xts_Q)
}
