% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maginature.R
\name{sig_Roll_mean}
\alias{sig_Roll_mean}
\title{Maxi / Mini value of roll windows}
\usage{
sig_Roll_mean(Data, roll_size = 7, max = TRUE)
}
\arguments{
\item{Data}{[xts / array]-2D(time, space) Time serious Data in difficult location.}

\item{roll_size}{[integer] Roll windows wedith.}

\item{max}{[logi] TRUE | FALSE for max or min.}
}
\value{
sig_Roll_mean [num] [1/step]
}
\description{
Maxi / Mini value of roll windows
}
\examples{
rslt <- sig_Roll_mean(xts_Q)
}
