% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{sig_Slope_fdc}
\alias{sig_Slope_fdc}
\title{Slope of the flow duration curve}
\usage{
sig_Slope_fdc(Data, tile_low = 33, tile_high = 66, log_fit = T)
}
\arguments{
\item{Data}{[xts / array]-2D(time, space) Time serious Data in difficult location.}

\item{tile_low}{[num]-range(0, 100) The lower tile, but the data is the bigger, because the FDC is in decreasing.}

\item{tile_high}{[num]-range(0, 100) The higher tile, but the data is the smaller, because the FDC is in decreasing.}

\item{log_fit}{|T| for mit ln(Q), |F| for ohne ln(Q)}
}
\value{
[array]-2D(1, space) The data in the tiles. Or [xts]-2D(sum_time, space) for other scale with apply.XXXly.
}
\description{
Slope of the flow duration curve of the data mostly streamflow.
For the other scale can use apply.yearly(Data, sig_Quantile, tile = c(5, 95)) or apply.weekly apply.monthly apply.quarterly apply.yearly.
}
\examples{
rslt <- sig_Slope_fdc(xts_Q)
rslt <- sig_Slope_fdc(xts_Q, 5, 50, FALSE)
rslt <- xts::apply.yearly(xts_Q, sig_Slope_fdc)
}
\references{
[Sawicz.2011]
}
