% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maginature.R
\name{sig_Snowday_ratio}
\alias{sig_Snowday_ratio}
\title{Snowday ratio}
\usage{
sig_Snowday_ratio(Preciptation, Temperatur, snow_threshold = 2)
}
\arguments{
\item{Preciptation}{[xts]-2D(time, space) Time serious Preciptation in difficult location.}

\item{Temperatur}{[xts]-2D(time, space) Time serious Temperatur in difficult location.}

\item{snow_threshold}{Temperatur threshold for the snow day}
}
\description{
the ration of the day with snow, with the Temperatur methods
}
\examples{
rslt <- sig_Snowday_ratio(xts_P, xts_Q)
}
