% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maginature.R
\name{sig_Stream_elas.Sawicz}
\alias{sig_Stream_elas.Sawicz}
\alias{sig_Stream_elas.Sanka}
\title{Streamflow-precipitation elasticity}
\usage{
sig_Stream_elas.Sawicz(
  Streamflow,
  Preciptation,
  t_scale = apply.yearly,
  wateryear_start_m = 1
)

sig_Stream_elas.Sanka(
  Streamflow,
  Preciptation,
  t_scale = apply.yearly,
  wateryear_start_m = 1
)
}
\arguments{
\item{Streamflow}{[xts]-2D(time, space) Time serious Streamflow in difficult location.}

\item{Preciptation}{[xts]-2D(time, space) Time serious Preciptation in difficult location.}

\item{t_scale}{[function] |apply.quarterly| for quarterly, |apply.yearly| for yearly.}

\item{wateryear_start_m}{[num]-|2|3|4|5|6|7|8|9|10|11|12| for Feb. ... Dec.}
}
\value{
stream elasticity [xts]-2D(time, space) in time scale
}
\description{
Streamflow-precipitation elasticity
}
\examples{
rslt <- sig_Stream_elas.Sawicz(xts_Q, xts_P)
rslt <- median(sig_Stream_elas.Sawicz(xts_Q, xts_P, wateryear_start_m = 10))
rslt <- sig_Stream_elas.Sanka(xts_Q, xts_P)
rslt <- median(sig_Stream_elas.Sanka(xts_Q, xts_P, wateryear_start_m = 10))
}
