% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{ts_interpolat_NA_time}
\alias{ts_interpolat_NA_time}
\title{linear interpolation}
\usage{
ts_interpolat_NA_time(Data)
}
\arguments{
\item{Data}{[xts]-2D(time, space) Time serious Data in difficult location.}
}
\value{
Interpolated time serious Data [xts]-2D(time, space)
}
\description{
linear interpolated with the last and next data (in time demision)
}
\examples{
xts_Q_NA <- xts_Q
xts_Q_NA[3:5,1] <- NA
rslt <- ts_interpolat_NA_time(xts_Q_NA)
}
