% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{ts_Roll_mean}
\alias{ts_Roll_mean}
\title{mean of roll windows}
\usage{
ts_Roll_mean(Data, roll_size = 7)
}
\arguments{
\item{Data}{[xts / array]-2D(time, space) Time serious Data in difficult location.}

\item{roll_size}{[integer] Roll windows wedith.}
}
\value{
new time series
}
\description{
mean of roll windows
}
\examples{
rslt <- ts_Roll_mean(xts_Q)
}
