% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{water_year}
\alias{water_year}
\title{Reshape the time series in water-year}
\usage{
water_year(Data, wateryear_start_m)
}
\arguments{
\item{Data}{[xts]-2D(time, space) Time serious Data in difficult location.}

\item{wateryear_start_m}{[num]-|2|3|4|5|6|7|8|9|10|11|12| for Feb. ... Dec.}
}
\value{
By water year reshaped time serious Data [xts]-2D(time, space)
}
\description{
Reshape the time series in water-year, set the start month of the water year,
the number of a water year will be same as the first year-number. And the data before the start month will in the end of the time series.
Please make sure the time series begin with the 1. Jan..
}
\examples{
water_year(xts_Q, 10)
}
