% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xts_Q}
\alias{xts_Q}
\title{Steamflow}
\format{
[xts] from 1999-01-01 to 2008-12-31 for two stations.

[xts] from 1999-01-01 to 2008-12-31 for two stations.
}
\usage{
xts_Q

xts_Q
}
\description{
Steamflow

Potential Evapotranspiration
}
\keyword{datasets}
