% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maginature.R
\name{sig_Runoff_ratio}
\alias{sig_Runoff_ratio}
\title{Runoff ratio}
\usage{
sig_Runoff_ratio(Streamflow, Preciptation)
}
\arguments{
\item{Streamflow}{[xts / array]-2D(time, space) Time serious Streamflow in difficult location.}

\item{Preciptation}{[xts / array]-2D(time, space) Time serious Preciptation in difficult location.}
}
\value{
[array]-2D(1, space) The Runoff ratio. Or [xts]-2D(sum_time, space) for other scale with apply.XXXly.
}
\description{
Runoff ratio of the time serise data.
For the other scale (if you want) can use apply.yearly or apply.weekly apply.monthly apply.quarterly apply.yearly see the example.
please make sure the Streamflow and Preciptation have the same unit.
}
\examples{
rslt <- sig_Runoff_ratio(xts_Q, xts_P)
rslt <- xts::apply.yearly(cbind(Q = xts_Q, P = xts_P), function(QP) sig_Runoff_ratio(QP[, grep("Q", colnames(QP))], QP[, grep("P", colnames(QP))]))
}
